refactor 'flow_control' to 'control_flow', removed pipe move, factorised move expressions in IfElse aspect
